http: only expect/continue when necessary
There's no need to do expect/continue generally, only when we're _not_
authenticated presently but _were_ authenticated with NTLM or Negotiate.
In this case, we'll re-use that authentication mechanism and we need to
use care in doing so to prevent POSTing data multiple times or losing
our POST data stream entirely.